fixed the data accessor problem with position angles. Unit test updated to ensure calling these new methods doesn't cause a crash (however there is no check that the result is sensible)
git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@1377 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Synthesis/synthesis/trunk/dataaccess/TableConstDataAccessor.h
@@ -221,10 +221,10 @@ private:
   CachedAccessorField<casa::Vector<casa::uInt> > itsFeed2;
   
   /// internal buffer for the position angles of the first feed
-  CachedAccessorField<casa::Vector<casa::Double> > itsFeed1PA;
+  CachedAccessorField<casa::Vector<casa::Float> > itsFeed1PA;
   
   /// internal buffer for the position angles of the second feed
-  CachedAccessorField<casa::Vector<casa::Double> > itsFeed2PA;
+  CachedAccessorField<casa::Vector<casa::Float> > itsFeed2PA;
   
   /// internal buffer for the pointing directions of the first antenna/feed
   CachedAccessorField<casa::Vector<casa::MVDirection> > itsPointingDir1;
